+++ frontend/constants.js
@@ -1,5 +1,6 @@
 import getConfig from './helpers/getConfig';
 
 export const { searchBarBlacklist: SEARCH_BAR_BLACKLIST } = getConfig();
+export const FILTER_BAR_PORTAL = 'filter-bar.content.before';
 
 export const CART_MAX_ITEMS = 99;