+++ frontend/portals/AppBarRight/connector.js
@@ -1,15 +1,19 @@
 import { connect } from 'react-redux';
 import { getCartProductDisplayCount } from '@shopgate/pwa-common-commerce/cart/selectors';
-import { isSearchBarVisible } from '../../selectors';
+import { makeGetIsAppBarIconVisible } from '../../selectors';
 
 /**
  * Maps the contents of the state to the component props.
  * @param {Object} state The current application state.
  * @return {Object} The extended component props.
  */
-const mapStateToProps = state => ({
-  isVisible: isSearchBarVisible(state),
-  count: getCartProductDisplayCount(state),
-});
+const makeMapStateToProps = () => {
+  const getIsAppBarIconVisible = makeGetIsAppBarIconVisible();
 
-export default connect(mapStateToProps);
+  return (state, props) => ({
+    isVisible: getIsAppBarIconVisible(state, props),
+    count: getCartProductDisplayCount(state),
+  });
+};
+
+export default connect(makeMapStateToProps);